radio: don't terminate transmission early
Previously, we were polling the TxFifoFull level, which indicates
whether the FIFO currently has data.

This is incorrect, because the fifo is considered "empty" as soon
as the last byte starts transmitting.

As a result, we were turning off the transmitter before the last
byte could be sent, and before the CRC16 could be transmitted.

Instead, wait for the "PacketSent" IRQ to go high.

Since we don't have this wired up, poll it in the register space.

Signed-off-by: Sean Cross <sean@xobs.io>